safari-600.7-branch/LayoutTests/http/tests/notifications/request-expected.txt

This test checks that a request for permission is made.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Requesting permission with default permission
PASS Notification.permission is "default"
PASS permission is granted
PASS Notification.permission is "granted"
Requesting permission with non-default permission
PASS permission is denied
PASS Notification.permission is "denied"
PASS successfullyParsed is true

TEST COMPLETE
